+++ assets/css/app.scss
@@ -1,5 +1,5 @@
 @import "sanitize.css";
 @import "base/background";
-@import "components/circle";
+@import "components/canvas";
 @import "components/player";
 @import "pages/signup";
