+++ assets/css/components/_canvas.scss
@@ -1,0 +1,5 @@
+#canvas {
+  width: 100vw;
+  height: 100vh;
+  outline: none;
+}
