--- assets/css/components/_circle.scss
@@ -1,13 +1,0 @@
-.circle {
-  border-radius: 50%;
-  border: solid 1px red;
-  position: absolute;
-  width: 100px;
-  height: 100px;
-}
-
-#canvas {
-  width: 100vw;
-  height: 100vh;
-  outline: none;
-}
