+++ assets/js/app.js
@@ -1,5 +1,8 @@
 import '../css/app.scss'
 import './shoelace'
+import { setupSpeaker } from './speaker'
 import { setupPlayer } from './spotify'
 
-setupPlayer()
+const { beat, changeLightColor } = setupSpeaker('canvas')
+
+setupPlayer({ onBeat: beat, onSection: changeLightColor })
