+++ lib/nyon_web/endpoint.ex
@@ -24,7 +24,7 @@ defmodule NyonWeb.Endpoint do
     at: "/",
     from: :nyon,
     gzip: false,
-    only: ~w(css fonts shoelace images js favicon.ico robots.txt)
+    only: ~w(css fonts shoelace images js models favicon.ico robots.txt)
 
   # Code reloading can be explicitly enabled under the
   # :code_reloader configuration of your endpoint.
