+++ lib/nyon_web/templates/music/player/index.html.eex
@@ -1,8 +1,5 @@
 <div id="token" data-token="<%= @account.access_token %>"></div>
 
-<div id="circle_beat" class="circle"></div>
-<div id="circle_bar" class="circle"></div>
-
 <canvas id="canvas"></canvas>
 
 <div class="player">
